[bockbuild] Fix bug with gdk-pixbuf-query-loaders
Solitary relocates some libraries in-place in the build-root, rendering
gdk-pixbuf-query-loaders unusuable after solitary completes. Since
gdk-pixbuf-query-loaders is required after the solitary step, this
commit introduces a little hack that changes the working dir to
$PREFIX/lib/ before running gdk-pixbuf-query-loaders. A real solution
instead of this hack would be patching solitary not to work in-place,
but relocate the libraries AFTER copying them to the bundle dir.